Link most of MetroTRK (free %%)

+++ include/TRK_MINNOW_DOLPHIN/ppc/Generic/mpc_7xx_603e.h
@@ -7,7 +7,7 @@
 extern "C" {
 #endif
 
-u32 TRKTargetCPUMinorType(void);
+u8 TRKTargetCPUMinorType(void);
 
 #ifdef __cplusplus
 }
