+++ src/TRK_MINNOW_DOLPHIN/debugger/embedded/MetroTRK/Processor/ppc/Generic/mpc_7xx_603e.c
@@ -245,4 +245,4 @@ lbl_8037149C:
 #endif // clang-format on
 }
 
-u32 TRKTargetCPUMinorType(void) { return 0x54; }
+u8 TRKTargetCPUMinorType(void) { return 0x54; }
